Change keybinding setup once again
So it turns out that, with default evdev config for inet symbols, two keycodes map to keysym XF86Send: 153 and 239
This causes issues when udev is configured to emit 239 for a key.
For example, if client software (such as a hotkey daemon) is only configured to recognise the keysym XF86Send, and it uses XKeysymToKeycode to discover the corresponding keycode, it will probably only listen for keycode 153.
This is very unexpected, so I changed the udev configuration to avoid this specific situation.