Add files to support a totally custom key-mapping setup
It appears trying to roll with already-provided configurations can cause a variety of issues.
I still think the current defaults more or less work for a quick setup, but something better is necessary.
I reintroduce the Advanced section to the README, as well as the `steelseries` file and more.
Hopefully this will ease the difficulty of configuring XKB for newcomers.